refactor: Rename WebSocketManagerImpl to ChatWebSocketDataSource and remove interface (#199)
- Renamed `WebSocketManagerImpl` to `ChatWebSocketDataSource` to better reflect its role as a data source for WebSocket communication.
- Removed the `WebSocketManager` interface and the corresponding `WebSocketModule` as `ChatWebSocketDataSource` is directly injected.
